# openapi.yaml
openapi: 3.0.3
info:
  title: Web service for Gift Certificates system
  #description: 
  #termsOfService: http://swagger.io/terms/
  #contact:
    #email: apiteam@swagger.io
  #license:
    #name: Apache 2.0
    #url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
#externalDocs:
  #description: Find out more about Swagger
  #url: http://swagger.io
servers:
  - url: https://example.org/api
tags:
  - name: tag
    description: CRD operations for tag
  - name: gift-certificate
    description: CRUD operations for gift certificate
paths:
  /tag:
    post:
      tags:
        - tag
      summary: Add a new tag
      description: add a new tag
      operationId: addTag
      requestBody:
        description: Add a new tag
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tag'
          application/xml:
            schema:
              $ref: '#/components/schemas/Tag'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Tag'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Tag'
        default:
        '405':
          description: invalid input
  /tags:
    get:
      tags:
        - tag
      summary: Find all tags
      description: find all tags
      operationId: getAllTags
      responses:
        default:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'          
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'
  /tag/{tagId}:
    get:
      tags:
        - tag
      summary: Find a tag by ID
      description: find a tag by ID
      operationId: getTagById
      parameters:
        - name: tagId
          in: path
          description: id of tag to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Tag'
        '400':
          description: invalid ID supplied
        default:
        '404':
          description: tag not found
    delete:
      tags:
        - tag
      summary: Delete a tag by ID
      description: delete a tag by ID
      operationId: deleteTag
      parameters:
        - name: tagId
          in: path
          description: Tag id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
        '405':
          description: invalid ID supplied
        default:
        '404':
          description: tag not found
  /gift-certificate:
    post:
      tags:
        - gift-certificate
      summary: Add a new gift certificate
      description: add a new gift certificate
      operationId: addGiftCertificate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GiftCertificate'
          application/xml:
            schema:
              $ref: '#/components/schemas/GiftCertificate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/GiftCertificate'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GiftCertificate'
        default:
        '405':
          description: invalid input
    put:
      tags:
        - gift-certificate
      summary: Update a gift certificate
      description: update a gift certificate
      operationId: updateGiftCertificate
      parameters:
        - name: id
          in: query
          description: id of the gift certificate that needs to be updated
          required: true
          schema:
            type: integer
            format: int64
        - name: name
          in: query
          description: name of the gift certificate that needs to be updated
          required: true
          schema:
            type: string
        - name: description
          in: query
          description: description of the gift certificate that needs to be updated
          required: false
          schema:
            type: string
        - name: price
          in: query
          description: price of the gift certificate that needs to be updated
          required: true
          schema:
            type: number
            format: double
        - name: duration
          in: query
          description: duration of the gift certificate that needs to be updated
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GiftCertificate'      
        '405':
          description: invalid parameters supplied
        default:
        '404':
          description: gift certificate not found
    get:
      tags:
        - gift-certificate
      summary: Find all gift certificates with specified parameters
      description:  find all gift certificates with specified parameters
      operationId: getAllGiftCertificatesWithParameters
      parameters:
        - name: tagName
          in: query
          description: gift certificate tag name that needs to be retrieved
          required: false
          schema:
            type: string
        - name: searchString
          in: query
          description: gift certificate part of name/description to be searched
          required: false
          schema:
            type: string
        - name: sortType
          in: query
          description: gift certificates sort type
          required: false
          schema:
            type: string
            enum: [date, name, dateThenName, nameThenDate]
        - name: sortOrder
          in: query
          description: gift certificates sort order
          required: false
          schema:
            type: string
            enum: [asc, desc]
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GiftCertificate'          
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GiftCertificate'
        default:
        '405':
          description: invalid parameters supplied
  /gift-certificates:
    get:
      tags:
        - gift-certificate
      summary: Find all gift certificates
      description: find all gift certificates
      operationId: getAllGiftCertificates
      responses:
        default:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GiftCertificate'          
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GiftCertificate'
  /gift-certificate/{gsId}:
    get:
      tags:
        - gift-certificate
      summary: Find gift certificate by ID
      description: Find gift certificate by ID
      operationId: getGiftCertificateById
      parameters:
        - name: gsId
          in: path
          description: ID of gift certificate that needs to be found
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GiftCertificate'          
            application/xml:
              schema:
                $ref: '#/components/schemas/GiftCertificate'
        '400':
          description: Invalid ID supplied
        default:
        '404':
          description: Gift certificate not found
    delete:
      tags:
        - gift-certificate
      summary: Delete gift certificate by ID
      description: Delete gift certificate by ID
      operationId: deleteGiftCertificate
      parameters:
        - name: gsId
          in: path
          description: ID of the gift certificate that needs to be deleted
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: Invalid ID supplied
        default:
        '404':
          description: gift certificate not found
components:
  schemas:
    Tag:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 198772
        name:
          type: string
          example: 'Birthday'
      xml:
        name: tag
    GiftCertificate:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 198772
        name:
          type: string
          example: 'Birthday certificate'
        description:
          type: string
          example: 'Some description'
        price:
          type: number
          format: double
          example: 1500.00
        duration:
          type: integer
          format: int64
          example: 94301
        create_date:
          type: string
          format: date-time
          example: '2017-07-21T17:32:28Z'
        last_update_date:
          type: string
          format: date-time
          example: '2017-07-21T17:32:28Z'
      xml:
        name: gift_certificate